mixoe/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "mixoe"
version = "0.1.0"
description = "Multi-task retinal image analysis with Mixture of Experts and Feature Pyramid Networks"
authors = [
    {name = "MixOE Team", email = "mixoe@example.com"},
]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "torch>=1.9.0",
    "torchvision>=0.10.0",
    "timm>=0.6.0",
    "opencv-python>=4.5.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "tqdm>=4.62.0",
    "albumentations>=1.1.0",
    "scikit-learn>=1.0.0",
    "PyYAML>=6.0",
    "wandb>=0.20.1",
]

[project.urls]
Homepage = "https://github.com/example/mixoe"
Repository = "https://github.com/example/mixoe"
Issues = "https://github.com/example/mixoe/issues"

[tool.setuptools]
py-modules = [
    "data",
    "network", 
    "loss",
    "trainer",
    "train",
    "config_reader",
]

[tool.setuptools.package-data]
"*" = ["config/*.yaml"] 
